adding async delete request with client side js

+++ course-total-progress/util/file.js
@@ -1,10 +1,14 @@
 const fs = require('fs');
 const deleteFile = (filePath) => {
-  fs.unlink(filePath, (err) => {
-    if (err) {
-      throw (err)
-    }
-  });
+  console.log(filePath);
+  
+  if (filePath) {
+    fs.unlink(filePath, (err) => {
+      if (err) {
+        throw (err)
+      }
+    });
+  }
 }
 
 exports.deleteFile = deleteFile;
